Move test dfs to DiGraph
